// src/screens/sign_In/__generated__/login.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: login
// ====================================================

export interface login_login {
  __typename: 'LoginResult'
  ok: boolean
  token: string | null
  error: string | null
}

export interface login {
  login: login_login
}

export interface loginVariables {
  email: string
  pwd: string
}